+++ src/FactoryInterface.php
@@ -3,6 +3,7 @@ declare(strict_types = 1);
 
 namespace Middlewares\Utils;
 
+use Psr\Http\Message\RequestFactoryInterface;
 use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ServerRequestFactoryInterface;
 use Psr\Http\Message\StreamFactoryInterface;
@@ -11,6 +12,8 @@ use Psr\Http\Message\UriFactoryInterface;
 
 interface FactoryInterface
 {
+    public function getRequestFactory(): RequestFactoryInterface;
+
     public function getResponseFactory(): ResponseFactoryInterface;
 
     public function getServerRequestFactory(): ServerRequestFactoryInterface;